// src/app/services/model/projectInformationRequestModel.ts
/**
 * Metal Casting Project API
 * This API serves the frontend with data from databases and other external services.
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProjectInformationRequestModel { 
    id?: number;
    projectNumber?: string | null;
    projectName?: string | null;
    isSequential?: boolean;
    projectDescription?: string | null;
    invoiceReferenceNumber?: string | null;
    companyRegistrationNumber?: string | null;
    assignedEmployees?: Array<string> | null;
    projectSatisfaction?: number;
    createdOn?: string;
}